resolution de conflit
